fix(desktop): route sidebar Conversations to the hub
Failure-Class: none
