+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-mainwindow.json
@@ -8,7 +8,7 @@
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/CanonicalMemoryAtlasView.swift": 4427,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/MemoryAtlasForceLayout.swift": 1646,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/TasksPage.swift": 5716,
-    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": 1690
+    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": 1627
   },
   "raise_justifications": {
     "desktop/macos/Desktop/Sources/MainWindow/Components/ChatBubble.swift": "Final-assistant Omi mark ownership and canonical expandable tool-call presentation are row-level rendering contracts that belong with bubble equality and layout. The mark remains an overlay so it cannot shift bubbles horizontally, while its owner row reserves the mark's height to prevent an empty streaming reply from clipping at the transcript live edge.",
