+++ desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift
@@ -2,78 +2,11 @@
 import OmiTheme
 import SwiftUI
 
-// MARK: - Navigation Item Model
-enum SidebarNavItem: Int, CaseIterable {
-  case dashboard = 0
-  case conversations = 1
-  case chat = 2
-  case memories = 3
-  case tasks = 4
-  case focus = 5
-  case insight = 6
-  case rewind = 7
-  case apps = 8
-  case settings = 9
-  case permissions = 10
-  case help = 12
-
-  var title: String {
-    switch self {
-    case .dashboard: return "Home"
-    case .conversations: return "Conversations"
-    case .chat: return "Chat"
-    case .memories: return "Memories"
-    case .tasks: return "Tasks"
-    case .focus: return "Focus"
-    case .insight: return "Insights"
-    case .rewind: return "Rewind"
-    case .apps: return "Apps"
-    case .settings: return "Settings"
-    case .permissions: return "Permissions"
-    case .help: return "Help from Founder"
-    }
-  }
-
-  var icon: String {
-    switch self {
-    case .dashboard: return "house.fill"
-    case .conversations: return "text.bubble.fill"
-    case .chat: return "bubble.left.and.bubble.right.fill"
-    case .memories: return "brain"
-    case .tasks: return "checklist"
-    case .focus: return "eye.fill"
-    case .insight: return "lightbulb.fill"
-    case .rewind: return "clock.arrow.circlepath"
-    case .apps: return "puzzlepiece.fill"
-    case .settings: return "gearshape.fill"
-    case .permissions: return "exclamationmark.triangle.fill"
-    case .help: return "bubble.left.fill"
-    }
-  }
-
-  /// Minimum tier level required to access this item (0 = always available)
-  var requiredTier: Int {
-    switch self {
-    case .conversations, .rewind: return 1
-    case .memories: return 2
-    case .tasks: return 3
-    case .chat: return 4
-    case .dashboard: return 5
-    case .apps: return 6
-    default: return 0
-    }
-  }
-
-  /// Items shown in the main navigation (top section)
-  static var mainItems: [SidebarNavItem] {
-    [.dashboard, .conversations, .memories, .tasks, .focus, .insight, .rewind, .apps]
-  }
-}
-
 // MARK: - Sidebar View
 struct SidebarView: View {
   @Binding var selectedIndex: Int
   @Binding var isCollapsed: Bool
+  @Binding var memoryDestinationRawValue: Int
   @ObservedObject var appState: AppState
   @ObservedObject private var authState = AuthState.shared
   @ObservedObject private var insightStorage = InsightStorage.shared
@@ -172,7 +105,11 @@ struct SidebarView: View {
                         }
                       }
                     }
-                    selectedIndex = item.rawValue
+                    MemoryHubDestination.applySidebarSelection(
+                      item,
+                      selectedIndex: &selectedIndex,
+                      memoryDestinationRawValue: &memoryDestinationRawValue
+                    )
                     AnalyticsManager.shared.tabChanged(tabName: item.title)
                   },
                   onToggle: {
