+++ desktop/macos/changelog/unreleased/20260731-sidebar-conversations-routing.json
@@ -1,0 +1,3 @@
+{
+  "change": "Fixed the Conversations sidebar item opening the Conversations hub"
+}
